Improved the DiamaeterOptimized function performance
Here the height of the tree is being calculated in the same recursion when the diameter is being calculated.
So complexity is O(n) instead of O(n^2) in the simple approach.
Also another thing noticed was in C program there were pointers to pass the height as a parameter along. However in Java we don;t have pointers,
so when the the Simple LinkedList Node was used as a pointer structure, the performance surprisingly degraded instead of improving. Hence the second parameter of the optimized function was changed to a single element array.Optimized. Here the height of the tree is being calculated in the same recursion when the diameter is being calculated.
So complexity is O(n) instead of O(n^2) in the simple approach.
Also another thing noticed was in C program there were pointers to pass the height as a parameter along. However in Java we don;t have pointers,
so when the the Simple LinkedList Node was used as a pointer structure, the performance surprisingly degraded instead of improving. Hence the second parameter of the optimized function was changed to a single element array.